+++ source/is/helper/isAsync.ts
@@ -17,4 +17,3 @@ import { isFunction } from './isFunction';
 export function isAsync<R>(value: R): value is Include<R, AsyncArrow> {
   return isFunction(value) && isEqual(value.constructor.name, ASYNC_FUNCTION);
 }
-