+++ source/is/index.ts
@@ -6,12 +6,25 @@
   ------------------------------------------------------------------------------
 */
 export { isArray } from './helper/isArray';
+export { isArrayLike } from './helper/isArrayLike';
+export { isAsync } from './helper/isAsync';
+export { isAsyncIterable } from './helper/isAsyncIterable';
 export { isBoolean } from './helper/isBoolean';
+export { isClient } from './helper/isClient';
+export { isDOM } from './helper/isDOM';
 export { isEqual } from './helper/isEqual';
+export { isFloat } from './helper/isFloat';
 export { isFunction } from './helper/isFunction';
 export { isInteger } from './helper/isInteger';
+export { isIterable } from './helper/isIterable';
+export { isLocalStorage } from './helper/isLocalStorage';
+export { isNaN } from './helper/isNaN';
 export { isNil } from './helper/isNil';
 export { isNull } from './helper/isNull';
 export { isNumber } from './helper/isNumber';
+export { isPromise } from './helper/isPromise';
+export { isRegExp } from './helper/isRegExp';
+export { isServer } from './helper/isServer';
 export { isString } from './helper/isString';
+export { isSymbol } from './helper/isSymbol';
 export { isUndefined } from './helper/isUndefined';