+++ source/type/collection/ArrayLikeLiteral.ts
@@ -1,0 +1,6 @@
+/**
+  Represents a type that behaves like an array by having a 'length' property.
+*/
+export interface ArrayLikeLiteral {
+  readonly length: number;
+}