+++ source/type/index.ts
@@ -6,6 +6,7 @@
   ------------------------------------------------------------------------------
 */
 
+export type { ArrayLikeLiteral } from './collection/ArrayLikeLiteral';
 export type { Arrow } from './collection/Arrow';
 export type { AsyncArrow } from './collection/AsyncArrow';
 export type { Include } from './collection/Include';